fix(vchart): rename file name
