+++ packages/vchart/src/series/index.ts
@@ -11,8 +11,8 @@ import type { ILineSeriesSpec } from './line/interface';
 import { BarSeries } from './bar/bar';
 import { Bar3dSeries } from './bar/bar-3d';
 import type { IBar3dSeriesSpec, IBarSeriesSpec } from './bar/interface';
-import { RangeColumnSeries } from './range-column/rangeColumn';
-import { RangeColumn3dSeries } from './range-column/rangeColumn-3d';
+import { RangeColumnSeries } from './range-column/range-column';
+import { RangeColumn3dSeries } from './range-column/range-column-3d';
 import type { IRangeColumn3dSeriesSpec, IRangeColumnSeriesSpec } from './range-column/interface';
 import { RangeAreaSeries } from './range-area/range-area';
 import type { IRangeAreaSeriesSpec } from './range-area/interface';
