+++ packages/vchart/src/series/range-column/range-column.ts
